requirements.txt
keras==2.11.0
numpy==1.21.5
mediapipe==0.9.0.1
opencv_contrib_python==4.6.0.66
Pillow==9.4.0
streamlit==1.19.0
tensorflow-cpu==2.11.0
